Adding Support for RHEL/CentOS 6&7

+++ Software/Client/Projects/ReseptConsoleClient/ktclient_remote_deploy.py
@@ -182,28 +182,54 @@ def deploy_site_config(ssh_host, config_path, installer_path, rccd_path):
                 ssh_host=quote(ssh_host),
                 temp_dir=quote(remote_temp_dir)))
 
-        run_remote_cmd(
-            ssh_host, """set -e;
-                         set -x;
-                         echo "Checking if apache2 is installed"
-                         which apache2
-                         (
-                             cd {temp_dir} &&
-                             tar xfz {installer_filename} &&
-                             (
-                               cd keytalkclient-* &&
-                                ./install.sh
-                             ) &&
-                             /usr/local/bin/keytalk/ktconfig --rccd-path {rccd_filename} &&
-                             cp {config_filename} /etc/keytalk/apache.ini &&
-                             /usr/local/bin/keytalk/renew_apache_ssl_cert --force &&
-                             echo "*  *  *  *  *   root    /usr/local/bin/keytalk/renew_apache_ssl_cert > $HOME/tmp/cron.ktapachecertrenewal.log 2>&1" > /etc/cron.d/keytalk
-                         ) &&
-                         rm -rf {temp_dir}""".format(
-                temp_dir=quote(remote_temp_dir), installer_filename=quote(
-                    os.path.basename(installer_path)), rccd_filename=quote(
-                    os.path.basename(rccd_path)), config_filename=quote(
-                        os.path.basename(config_path))), only_stdout=True)
+	if(run_cmd('lsb_release --id --short') == "Debian" or run_cmd('lsb_release --id --short') == "Ubuntu"):
+		run_remote_cmd(
+		    ssh_host, """set -e;
+		                 set -x;
+		                 echo "Checking if apache2 is installed"
+		                 which apache2
+		                 (
+		                     cd {temp_dir} &&
+		                     tar xfz {installer_filename} &&
+		                     (
+		                       cd keytalkclient-* &&
+		                        ./install.sh
+		                     ) &&
+		                     /usr/local/bin/keytalk/ktconfig --rccd-path {rccd_filename} &&
+		                     cp {config_filename} /etc/keytalk/apache.ini &&
+		                     /usr/local/bin/keytalk/renew_apache_ssl_cert --force &&
+		                     echo "*  *  *  *  *   root    /usr/local/bin/keytalk/renew_apache_ssl_cert > $HOME/tmp/cron.ktapachecertrenewal.log 2>&1" > /etc/cron.d/keytalk
+		                 ) &&
+		                 rm -rf {temp_dir}""".format(
+		        temp_dir=quote(remote_temp_dir), installer_filename=quote(
+		            os.path.basename(installer_path)), rccd_filename=quote(
+		            os.path.basename(rccd_path)), config_filename=quote(
+		                os.path.basename(config_path))), only_stdout=True)
+
+	if(run_cmd('lsb_release --id --short') == "RedHatEnterpriseServer" or run_cmd('lsb_release --id --short') == "CentOS"):
+		run_remote_cmd(
+		    ssh_host, """set -e;
+		                 set -x;
+		                 echo "Checking if httpd is installed"
+		                 which httpd
+		                 (
+		                     cd {temp_dir} &&
+		                     tar xfz {installer_filename} &&
+		                     (
+		                       cd keytalkclient-* &&
+		                        ./install.sh
+		                     ) &&
+		                     /usr/local/bin/keytalk/ktconfig --rccd-path {rccd_filename} &&
+		                     cp {config_filename} /etc/keytalk/apache.ini &&
+		                     /usr/local/bin/keytalk/renew_apache_ssl_cert --force &&
+		                     echo "*  *  *  *  *   root    /usr/local/bin/keytalk/renew_apache_ssl_cert > $HOME/tmp/cron.ktapachecertrenewal.log 2>&1" > /etc/cron.d/keytalk
+		                 ) &&
+		                 rm -rf {temp_dir}""".format(
+		        temp_dir=quote(remote_temp_dir), installer_filename=quote(
+		            os.path.basename(installer_path)), rccd_filename=quote(
+		            os.path.basename(rccd_path)), config_filename=quote(
+		                os.path.basename(config_path))), only_stdout=True)
+
     except CmdFailedException as ex:
         return ex.format_indented_message('Could not deploy to {}:'.format(ssh_host))
     return None
